Added the DTMF handler task, yet empty

+++ ccs_workspace/sandbox/dsp/dsp_goertzel.h
@@ -8,6 +8,6 @@
 #ifndef DSP_DSP_GOERTZEL_H_
 #define DSP_DSP_GOERTZEL_H_
 
-double goertzel(Uint32*, Uint16, Uint32, Uint32);
+float goertzel(Int16*, Uint16, Uint32, Uint32);
 
 #endif /* DSP_DSP_GOERTZEL_H_ */
